pyproject.toml
[tool.poetry]
name = "capstoneproject"
version = "0.1.0"
description = ""
authors = ["ngovandong <dongngo2001@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.1.5"
djangorestframework = "^3.14.0"
django-debug-toolbar = "^3.8.1"
djangorestframework-simplejwt = "^5.2.2"
django-cors-headers = "^3.13.0"
python-dotenv = "^0.21.1"
requests = "^2.28.2"
pillow = "^9.4.0"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
